Data used to model the DESPLIM compactness metric

Clean data used to model the DESPLIM compactness metric, based on the data provided by Kaufman et al. (2021). The metrics have been computed using the redistmetrics package.

Usage

compact_train

Format

A data frame with 547 rows and 11 columns:

compact

Kaufman compactness

boyce

Boyce Clark ratio

box_reock

Box Reock compactness

hull

Convex Hull compactness

len_wid

Length-Width compactness

polsby

Polsby Popper compactness

reock

Reock compactness

schwartz

Schwartzberg compactness

skew

Skew compactness

sym_x

X symmetry

sym_y

Y symmetry

Source

https://dataverse.harvard.edu/file.xhtml?fileId=4143644&version=1.1

https://github.com/aaronrkaufman/compactness/tree/e91519e15fd3ca55f57fc21323bb12ed5a66305b/data

https://alarm-redist.org/redistmetrics/