Calculate the 'desplim' compactness of an sf POLYGON object

Based on the approach described in Kaufmann et al. (2018), this function computes the compactness of an sf POLYGON object.

Usage

desplim_compactness(input, keep_metrics = FALSE)

Arguments

input

object of class sf of type POLYGON or MULTIPOLYGON.

keep_metrics

logical; should all the compactness metrics be kept? Default is FALSE.

Value

An sf object of type POLYGON matching input, with the compactness metric(s) added.

Details

Please see the dedicated vignette for information about the compactness metric. If input contains geometries of type MULTIPOLYGON, they are cast to POLYGON before computing.