Deterministically merge a set of polygons with possible paramters defined by the user. #' @description Function for merging a set of polygons, with optinal rules defined by the user. The merging process is deterministic, leading to reproducible results on subsequent runs with the same input data and rules.

Deterministically merge a set of polygons with possible paramters defined by the user. #' @description Function for merging a set of polygons, with optinal rules defined by the user. The merging process is deterministic, leading to reproducible results on subsequent runs with the same input data and rules.

Usage

desplim_merge(
  input_polygons,
  input_buildings = NULL,
  compact_method = "desplim",
  compact_threshold = 0.7,
  compact_allow = 0.2,
  area_threshold = 10000,
  enclosed_threshold = 0.8,
  building_threshold = 5
)

Arguments

input_polygons

object of class sf of type POLYGON to be merged.

input_buildings

object of class sf of type POLYGON or MULTIPOLYGON. An optional sf object representing buildings which should be considered when merging. Default is NULL.

compact_method

string; chosen method for calculating compactness of polygons. Default is "desplim", using the internal desplim_compactness function. Other options include "polsby", "schwartz" and "convex_hull" from the redistmetrics package.

compact_threshold

numerical; the baseline expected compactness of the output polygons. Default is 0.7.

compact_allow

numerical; the fraction of polygons allowed to fall below compact_tolerance. Default is 0.2.

area_threshold

numerical; the minimum required area of merged polygons. Default is 10000 units.

enclosed_threshold

numerical; the fraction of a polygon border shared with a single neighbour for it to be considered enclosed (and flagged for merging). Default is 0.8.

building_threshold

integer; the minimum number of required buildings within merged polygons if input_buildings is provided. Default is 5, but sis set to 0 if input_buildings is not provided.

Value

An sf object of POLYGONs resulting from the deterministic merge.

Details

There are no requirements for all polygons being adjacent to each other. Althugh named input_buildings, any sf object of type POLYGON or MULTIPOLYGON can be provided for this argument. For details about the compactness metric, please see the dedicated vignette. It is not recommended to set compact_allow too close to 1 and compact_tolerance too low. This is especially true if compact_method is set to "polsby".