Perform a detailed split of a polygon based on a sf LINESTRING object and optional border connections.

Perform a detailed split of a polygon based on a sf LINESTRING object and optional border connections.

Usage

desplim_split(
  input_polygon,
  input_lines,
  input_buildings = NULL,
  enable_line_nearest_node = TRUE,
  enable_border_connect = FALSE,
  ...
)

Arguments

input_polygon

object of class sf of type POLYGON to be split.

input_lines

object of class sf of type LINESTRING or MULTILINESTRING to be used for splitting.

input_buildings

object of class sf of type POLYGON or MULTIPOLYGON. An optional sf object representing buildings which cannot be intersected during the splitting procedure. Default is NULL.

enable_line_nearest_node

logical; whether to connect leaf nodes of input_lines to the nearest node using desplim_line_nearest_node. Default is TRUE.

enable_border_connect

logical; whether to connect input_lines to the border of input_polygon using desplim_connect_border. Default is FALSE.

...

additional arguments passed to desplim_connect_border if set to TRUE.

Value

An sf object of POLYGONs resulting from the split.

Details

The function splits an input polygon based on an initial set of linestrings. If enable_line_nearest_node is TRUE, leaf nodes in the input linestrings are connected to the nearest node as laid out in desplim_line_nearest_node. If enable_border_connect is TRUE, the input linestrings are connected to the border of the input polygon using desplim_connect_border.